Fix CFInfo layout on big endian platforms.
Prior to this change 64-bit platforms treated the _cfinfoa field
as both a uint64_t and a uint32_t. This works on little endian
platforms, however on big endian platforms the low 32 bits of a
uint64_t value are located at a four byte offset from the variable.

Fix this by always accessing the field as a uint64_t when on a
64-bit platform.